just some updates to make the function safe

+++ Graph/graph_adjList.cpp
@@ -15,10 +15,10 @@ void add_vertices(vector<long> vertex_list[], long source, long destination){
 }
 
 void print_graph(vector<long> graph[]){
-  long no_of_vertices = graph[1].size();
-  for(int i=0; i < no_of_vertices; i++){
+  unsigned long no_of_vertices = graph[1].size();
+  for(unsigned long i=0; i < no_of_vertices; i++){
     cout<<"["<<i<<"]";
-    for(int j = 0; j< graph[i].size(); j++){
+    for(unsigned long j = 0; j< graph[i].size(); j++){
       cout<<"--> ("<<graph[i][j]<<")";
     }
     cout<<"\n";
